moves the files and fixes the references
