+++ src/signals/incident/components/IncidentPreview/components/ContainerListPreview/ContainerListPreview.test.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import type { Container } from 'shared/types/incident';
 
-import ContainerList from '../../../form/ContainerSelect/ContainerList';
+import ContainerList from '../../../form/ContainerSelect/components/ContainerList';
 
 import ContainerListPreview from './ContainerListPreview';
 
-jest.mock('../../../form/ContainerSelect/ContainerList', () => jest.fn().mockImplementation(() => null));
+jest.mock('signals/incident/components/form/ContainerSelect/components/ContainerList', () => jest.fn().mockImplementation(() => null));
 
 describe('ContainerListPreview', () => {
   it('should render ContainerList with props', () => {
