--- src/signals/incident/components/form/ContainerSelect/LegendPanel/index.test.tsx
@@ -1,34 +1,0 @@
-import { render, screen } from '@testing-library/react';
-import React from 'react';
-import { withAppContext } from 'test/utils';
-
-import LegendPanel from '.';
-import type { LegendPanelProps } from '.';
-
-describe('LegendPanel', () => {
-  const props: LegendPanelProps = {
-    onClose: () => {},
-    title: 'Title',
-    variant: 'drawer',
-    items: [
-      {
-        iconUrl: 'url',
-        id: 'id',
-        label: 'label',
-      },
-    ],
-  };
-  it('should render', () => {
-    render(withAppContext(<LegendPanel {...props} />));
-
-    expect(screen.getByRole('list')).toBeInTheDocument();
-    props.items?.forEach(({ id }) => expect(screen.getByTestId(`legend-icons-item-${id}`)));
-    expect(screen.getAllByRole('listitem').length).toBe(props.items?.length);
-  });
-
-  it('should render with empty items', () => {
-    render(withAppContext(<LegendPanel onClose={props.onClose} title={props.title} variant={props.variant} />));
-
-    expect(screen.queryAllByRole('listitem').length).toBe(0);
-  });
-});
