+++ src/signals/incident/components/form/ContainerSelect/components/ContainerList/index.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import IconList from 'components/IconList/IconList';
+import type { Item } from 'signals/incident/components/form/ContainerSelect/types';
+
+export interface ContainerListProps {
+  selection: Item[];
+  size?: number;
+  className?: string;
+}
+
+const ContainerList: React.FC<ContainerListProps> = ({ selection, size, className }) =>
+  <IconList
+    className={className}
+    id="containerList"
+    size={size}
+    items={selection.map(({ id, description, iconUrl }) => ({ iconUrl, label: `${description} - ${id}`, id }))}
+  />;
+
+export default ContainerList;
