+++ src/signals/incident/components/form/ContainerSelect/components/Selector/components/LegendPanel/index.tsx
