+++ src/signals/incident/components/form/ContainerSelect/components/Selector/components/LegendToggleButton/index.tsx
